Refactor mcp service implementation
