+++ mcp/main.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	"mcp-server/pkg/client"
 	mcp "mcp-server/pkg/mcp"
 	"mcp-server/pkg/service"
 )
@@ -128,7 +127,7 @@ func serveRequest(c *gin.Context) {
 func main() {
 	router := service.GetRouter()
 	router.POST("/mcp", serveRequest)
-	router.POST("/client", client.ServeThirdPartyRequest)
+	router.POST("/client", mcp.ServeThirdPartyRequest)
 	cfg, err := service.InitConfig()
 	if err != nil {
 		logger.Error("Failed to get configurations", "error", err)
